feat
Share functionality now completely works and adds the video to the database. As of right now there are a few little things I need to implement to make the share functionality perfect. These include: reworking how youtube titles are added to the database, working with both edit texts, fixing the search functionality for the second edit text. Otherwise the share functionality is now done. To be revisited later