feat
fixed the way the viewpager works such that the positions of tabs are correct and the content provider is correctly loading the data. Also moved MainActivity's call to start VideoSYncing to onStart(). This is because navigating from other activities back to the MainActivity should start the triggering syncing the default video data.