fix: some major fixes, and doc changes
Fixed an issue happening with widget that caused the app to crash on some devices. Removed proguard because apparently its stopping some devices from building and I actually have no idea what it was doing in the first place. Moved some hard coded strings to the strings file in res. Added some stuff to the TODO folder (such as moving some of the widget video play code to the widget class instead of the videosyncintentservice)